requirements.txt
pyfiglet
colorama
langchain
langchain_core
langchain_openai
langchain_google_genai
pandas
python-dotenv
inquirer